+++ openfisca_uk/variables/finance/benefit/benefit.py
@@ -73,6 +73,7 @@ class household_benefits(Variable):
     def formula(household, period, parameters):
         HOUSEHOLD_BENEFITS = [
             "winter_fuel_allowance",
+            "energy_bills_rebate",
         ]
         personal_benefits = aggr(household, period, ["benefits"])
         household_benefits = add(household, period, HOUSEHOLD_BENEFITS)
@@ -250,6 +251,7 @@ class personal_benefits(Variable):
             "SSP",
             "SMP",
             "ssmg",
+            "basic_income",
         ]
         return add(person, period, BENEFITS)
 
@@ -292,15 +294,22 @@ class claims_all_entitled_benefits(Variable):
 
     def formula(benunit, period, parameters):
         # Return false we have any reported values in the simulation for benefits.
-        return aggr(benunit, period, [
-            "child_tax_credit_reported",
-            "working_tax_credit_reported",
-            "universal_credit_reported",
-            "housing_benefit_reported",
-            "JSA_income_reported",
-            "income_support_reported",
-            "ESA_income_reported",
-        ]).sum() < 1
+        return (
+            aggr(
+                benunit,
+                period,
+                [
+                    "child_tax_credit_reported",
+                    "working_tax_credit_reported",
+                    "universal_credit_reported",
+                    "housing_benefit_reported",
+                    "JSA_income_reported",
+                    "income_support_reported",
+                    "ESA_income_reported",
+                ],
+            ).sum()
+            < 1
+        )
 
 
 class claims_legacy_benefits(Variable):