+++ CHANGELOG.md
@@ -5,7 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.11.0] - 2022-01-01
+## [0.12.3] - 2022-02-14
+
+### Changed
+
+- Set basic income phaseout threshold default to 0.
+
+## [0.12.2] - 2022-02-14
+
+### Added
+
+- Basic income parameters and logic.
+
+## [0.12.1] - 2022-02-14
+
+### Changed
+
+* OpenFisca-Tools bumped to v0.3
+
+## [0.12.0] - 2022-02-06
+
+### Added
+
+* The Energy Bills Rebate scheme.
+
+## [0.11.0] - 2022-02-04
 
 ## Added